requirements.txt
requests
rich
python-dotenv
patch-ng
